edit Rose.java logic to send responses as a string to the gui
