+++ data/Rose.txt
@@ -1,4 +1,4 @@
-T | 0 | 2103 ip
 D | 0 | meow | 2025-01-26 1800
 E | 0 | birthday | tonight | tomorrow
 T | 0 | say hello
+T | 1 | buy valentines gift
