+++ src/main/java/rose/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package rose.ui.gui;
+package rose.gui;
 
 import java.io.IOException;
 import java.util.Collections;
