+++ src/main/java/rose/gui/MainWindow.java
@@ -1,4 +1,4 @@
-package rose.ui.gui;
+package rose.gui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -6,7 +6,9 @@ import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.*;
-import rose.ui.Rose;
+import rose.Rose;
+import rose.exceptions.RoseException;
+
 /**
  * Controller for the main GUI.
  */
@@ -38,13 +40,17 @@ public class MainWindow extends AnchorPane {
      * the dialog container. Clears the user input after processing.
      */
     @FXML
-    private void handleUserInput() {
+    private void handleUserInput() throws RoseException {
         String input = userInput.getText();
-        String response = rose.getResponse(input);
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, userImage),
-                DialogBox.getDukeDialog(response, dukeImage)
-        );
-        userInput.clear();
+        try {
+            String response = rose.getResponse(input);
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getUserDialog(input, userImage),
+                    DialogBox.getDukeDialog(response, dukeImage)
+            );
+            userInput.clear();
+        } catch (Exception e) {
+            throw new RoseException(e.getMessage());
+        }
     }
 }
