+++ src/main/java/rose/storage/Storage.java
@@ -1,4 +1,10 @@
-package rose.ui;
+package rose.storage;
+import rose.tasks.Deadline;
+import rose.tasks.Event;
+import rose.tasks.Todo;
+import rose.exceptions.RoseException;
+import rose.tasks.Task;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
