+++ src/main/java/rose/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package rose.ui;
+package rose.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
