+++ src/main/java/rose/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package rose.ui;
+package rose.tasks;
 
 import java.util.ArrayList;
