+++ src/test/java/rose/tasks/TaskListTest.java
@@ -1,9 +1,10 @@
-package rose.ui;
+package rose.tasks;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
+import rose.tasks.Task;
+import rose.tasks.TaskList;
+import rose.tasks.Todo;
 
 import static org.junit.jupiter.api.Assertions.*;
